BLUE GREEN GROOVY SCRIPT WORKS! This is for the pattern of common, a, b dirs

+++ blue_green/src/main/java/com/example/blue_green/blueGreen.groovy
@@ -9,7 +9,6 @@ logger.info('Hello from a Job DSL script!')
 
 /**
  * STAGE1 deploy_green
- * get_outputs()
  * get_blue()
  * change attach_asg_to var, apply 
  * deploy green() // wait for green to respond with 200s
@@ -26,30 +25,6 @@ logger.info('Hello from a Job DSL script!')
 // make this into a map value 1 = terragrunt_working_dir
 // value 2 = test location
 def String terragrunt_working_dir = '/Users/bskeen/repository/appeals-terraform/live/uat/revproxy-caseflow-replica'
-
-public Map get_outputs(terragrunt_working_dir) {
-	logger.info('Running get_outputs()')
-	def String infra_set = 'common'
-	def jsonSlurper = new JsonSlurper()
-  	def init_sout = new StringBuilder(), init_serr = new StringBuilder()
-	def proc_init =	"terragrunt init --terragrunt-source-update --terragrunt-working-dir ${terragrunt_working_dir}/${infra_set}".execute()
-    proc_init.consumeProcessOutput(init_sout, init_serr) 
-    proc_init.waitForOrKill(9000000)
-	
-	def sout = new StringBuilder(), serr = new StringBuilder()
-  	def proc = "terragrunt output -json --terragrunt-source-update --terragrunt-working-dir ${terragrunt_working_dir}/${infra_set}".execute() 
-  	proc.consumeProcessOutput(sout, serr) 
-  	proc.waitForOrKill(9000000)
-  	def object = jsonSlurper.parseText(sout.toString()) 
-  	def Map outputs = [
-	'attach_asg_to':object.get('attach_asg_to').get('value'),
-  'blue_weight_a':object.get('blue_weight_a').get('value'), 
-	'blue_weight_b':object.get('blue_weight_b').get('value'),
-	'green_weight_a':object.get('green_weight_a').get('value'),
-	'green_weight_b':object.get('green_weight_b').get('value'),
-  	]
-	return outputs
-}
 
 public def tg_apply(terragrunt_working_dir, infra_set) {
 	logger.info('Running tg_apply()')
@@ -71,18 +46,36 @@ public def tg_destroy(terragrunt_working_dir, infra_set) {
 	logger.info("PROC_APPLY SOUT = ${apply_sout}")
 }
 
-// not sure if this is needed but might come in handly later
 public def get_blue_green(terragrunt_working_dir) {	
 	logger.info("Running get_blue_green()")
-	def Map outputs = get_outputs(terragrunt_working_dir)
-	if (outputs.get('blue_weight_a').equals(100)) {
+	def String infra_set = 'common'
+	def jsonSlurper = new JsonSlurper()
+	def init_sout = new StringBuilder(), init_serr = new StringBuilder()
+	def proc_init =	"terragrunt init --terragrunt-source-update --terragrunt-working-dir ${terragrunt_working_dir}/${infra_set}".execute()
+	proc_init.consumeProcessOutput(init_sout, init_serr) 
+	proc_init.waitForOrKill(9000000)
+	
+	def sout = new StringBuilder(), serr = new StringBuilder()
+	def proc = "terragrunt output -json --terragrunt-source-update --terragrunt-working-dir ${terragrunt_working_dir}/${infra_set}".execute() 
+	proc.consumeProcessOutput(sout, serr) 
+	proc.waitForOrKill(9000000)
+	def object = jsonSlurper.parseText(sout.toString()) 
+	def Map outputs = [
+	'attach_asg_to':object.get('attach_asg_to').get('value'),
+	'blue_weight_a':object.get('blue_weight_a').get('value'), 
+	'blue_weight_b':object.get('blue_weight_b').get('value'),
+	'green_weight_a':object.get('green_weight_a').get('value'),
+	'green_weight_b':object.get('green_weight_b').get('value'),
+	]
+	if (outputs.get('blue_weight_a') >= 50) {
 		blue = 'a'
 	}
-	else if (outputs.get('blue_weight_b').equals(100)) {
+	else if (outputs.get('blue_weight_b') >= 50) {
 		blue = 'b'
 	} 
 	else {
-		logger.info("ERROR: Neither blue_weight_a or blue_weight_b is set to 100")
+		logger.error("ERROR: Neither blue_weight_a or blue_weight_b is greater than 50")
+		System.exit(1)	
 	}
 
 	if (outputs.get('green_weight_a').equals(100)) {
@@ -92,17 +85,16 @@ public def get_blue_green(terragrunt_working_dir) {
 		green = 'b'
 	} 
 	else {
-		logger.info("ERROR: Neither green_weight_a or green_weight_b is set to 100")
+		logger.error("ERROR: Neither green_weight_a or green_weight_b is set to 100")
+		System.exit(1)
 	}
-	logger.info("BLUE = ${blue}")
-	logger.info("GREEN = ${green}")
-	return [blue, green]
+	logger.info("OUTPUTS = ${outputs}")
+	return [blue, green, outputs]
 }
 
 public def change_attach_asg_to(terragrunt_working_dir) {
-	def Map outputs = get_outputs(terragrunt_working_dir)
 	logger.info("Running change_attach_asg_to()")
-	(blue, green) = get_blue_green(terragrunt_working_dir)	
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)	
 	if (blue.compareTo('a').equals(0)) {
 		attach_asg_to = 'b'
 		logger.info("ATTACHING TO ${attach_asg_to}")
@@ -127,32 +119,35 @@ public def change_attach_asg_to(terragrunt_working_dir) {
 
 public def deploy_green(terragrunt_working_dir) {
 	logger.info('Running deploy_green()')
-	(blue, green) = get_blue_green(terragrunt_working_dir)
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)
 	logger.info("DEPLOYING ${green}")
 	tg_apply(terragrunt_working_dir, green)
 }
 
 public def weight_shift(terragrunt_working_dir) {
 	logger.info('Running weight_shift()')
-	def String infra_set = 'common'
-	(blue, green) = get_blue_green(terragrunt_working_dir)
-	def Map outputs = get_outputs(terragrunt_working_dir)
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)
+	def infra_set = 'common'
 	
-	Integer x = 0
 	Integer blue_weight_a = outputs.get('blue_weight_a')
 	Integer blue_weight_b = outputs.get('blue_weight_b')
-
-	while (x.compareTo(100).equals(-1)) {
+	
+	while (blue_weight_a != 100 || blue_weight_b != 100) {
 		// blue weight shift starts here
 		if (blue.compareTo('a').equals(0)) {
 			blue_weight_a = blue_weight_a-10
 			blue_weight_b = blue_weight_b+10
 		}	
+		
 		else if (blue.compareTo('b').equals(0)) {
 			blue_weight_a = blue_weight_a+10
 			blue_weight_b = blue_weight_b-10
 		}	
 		
+		if (blue_weight_a == 110 || blue_weight_b == 110) {		
+        	break	
+        }
+
 		File tfvars = new File("${terragrunt_working_dir}/${infra_set}/terraform.tfvars")
 		if (tfvars.canRead()) {
 			tfvars.delete()
@@ -164,14 +159,31 @@ public def weight_shift(terragrunt_working_dir) {
 		tfvars.append "green_weight_b = ${outputs.get('green_weight_b')}\n"
 		tg_apply(terragrunt_working_dir, infra_set) // only changes the attach_asg_to var in common
 		tfvars.delete()
-		x = x+10
 		sleep(10000)// sleeps for 10s
 	}
 }
 
+public def custom_blue_weights(terragrunt_working_dir, blue_custom_weight_a, blue_custom_weight_b) {
+	logger.info('Running custom_weights')
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)
+	String infra_set = 'common'
+
+	File tfvars = new File("${terragrunt_working_dir}/${infra_set}/terraform.tfvars")
+	if (tfvars.canRead()) {
+		tfvars.delete()
+	}
+	tfvars.append "attach_asg_to = \"${outputs.get('attach_asg_to')}\"\n"
+	tfvars.append "blue_weight_a = ${blue_custom_weight_a}\n"
+	tfvars.append "blue_weight_b = ${blue_custom_weight_b}\n"
+	tfvars.append "green_weight_a = ${outputs.get('green_weight_a')}\n"
+	tfvars.append "green_weight_b = ${outputs.get('green_weight_b')}\n"
+	tg_apply(terragrunt_working_dir, infra_set) // only changes the attach_asg_to var in common
+	tfvars.delete()
+}
+
 public def destroy_old_blue(terragrunt_working_dir) {
 	logger.info("Running destroy_old_blue()")
-	(blue, green) = get_blue_green(terragrunt_working_dir)
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)
 	if (blue.compareTo('a').equals(0)) {
 		old_blue = 'b'
 	}
@@ -184,21 +196,21 @@ public def destroy_old_blue(terragrunt_working_dir) {
 
 public def destroy_green(terragrunt_working_dir) {
 	logger.info("Running destroy_green()")
-	(blue, green) = get_blue_green(terragrunt_working_dir)
+	(blue, green, outputs) = get_blue_green(terragrunt_working_dir)
 	if (green.compareTo('a').equals(0)) {
 		green = 'a'
 	}
 	else if (blue.compareTo('b').equals(0)) {
     	green = 'a'
     }
-	logger.info("DESTROYING OLD BLUE ${old_blue}")
-	tg_destroy(terragrunt_working_dir, old_blue)
+	logger.info("DESTROYING OLD BLUE ${green}")
+	tg_destroy(terragrunt_working_dir, green)
 }
 
 public def update_green(terragrunt_working_dir) {
   logger.info("Running update_green()")
   def String infra_set = 'common'
-  (blue, green) = get_blue_green(terragrunt_working_dir)
+  (blue, green, outputs) = get_blue_green(terragrunt_working_dir)
   if (blue.compareTo('a').equals(0)) {
   	green_weight_a = 0 
     green_weight_b = 100
@@ -207,7 +219,6 @@ public def update_green(terragrunt_working_dir) {
     green_weight_a = 100 
     green_weight_b = 0
   }
-  def Map outputs = get_outputs(terragrunt_working_dir)
   File tfvars = new File("${terragrunt_working_dir}/${infra_set}/terraform.tfvars")
   if (tfvars.canRead()) {
   	tfvars.delete()
@@ -222,14 +233,15 @@ public def update_green(terragrunt_working_dir) {
 }
 
 
-
-
 // Treat here and down as main()
 // Jenkins pipeline would pass around vars in / out instead of this file 
 logger.info("Starting...")
 change_attach_asg_to(terragrunt_working_dir)
 deploy_green(terragrunt_working_dir)
-
+// def blue_custom_weight_a = 100 
+// def blue_custom_weight_b = 0 
+// custom_blue_weights(terragrunt_working_dir, blue_custom_weight_a, blue_custom_weight_b) 
+// run_tests() // TODO: figure this out. Probably a script that runs - discuss it with team
 weight_shift(terragrunt_working_dir)
 destroy_old_blue(terragrunt_working_dir)
 update_green(terragrunt_working_dir)
